requirements.txt
Flask>=3.0
Flask-SQLAlchemy>=3.1
Flask-Migrate>=4.0
Flask-Login>=0.6
Flask-WTF>=1.2
email-validator>=2.1
Flask-Limiter>=3.7
python-dotenv>=1.0
Pillow>=10.0
psycopg2-binary>=2.9
passlib[bcrypt]>=1.7
bcrypt==4.0.1
